usb: dwc3: Drive a pulse on DP on CDP detection
Some USB host PCs switch the downstream port mode from
CDP to SDP if the device does not pull DP up within certain
duration (~2s).
This limits the current drawn by the device to 500mA(or 900mA)
as opposed to the possible 1.5A resulting in slow charging of
the battery.
Fix this by driving a DP pulse when there is a connect
notification from PMIC and detected the charger type as CDP.
Incase of nopmi devices, charger detection support present
in synopsys phy will perform charger detection. If CDP is detected
then driving of DP pulse can be done.

Change-Id: I7655a0812ba85bdc29fc4f66d4675717f5b4cd8f
Signed-off-by: Elson Roy Serrao <eserrao@codeaurora.org>
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>