usb: phy: Add support for PHY based charger detection for snps phy
The PHY IP supports different circuit blocks like voltage
sources, current sources, comparators and resistances on the
DP/DM lines with the control by registers. Add driver support
for detecting SDP, CDP and DCP chargers with the BC1.2 protocol
and notifying the same to the charger driver. Also, notify the
USB connect to the USB driver in case of SDP and CDP to support
enumeration.

Change-Id: Icee0d4634d78b6d278220443dd3cf502f54ff148
Signed-off-by: Rohith Kollalsi <quic_rkollals@quicinc.com>
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>