ep_pcie: core: Initialize the edma only after successful enumeration
If enumeration fails all the resources will be turned off and then
if try to access any register after turnng off the resources the system
is crashing due to NOC error for unclocked access.

In edma init function, the driver is trying to access DBI space which is
resulting to this NOC error as clocks are turned off already.

Change-Id: I3b0a5c02df84671dc96bcb6e99de2ff4ea1ee069
Signed-off-by: Krishna chaitanya chundru <quic_krichai@quicinc.com>