usb: gadget: Add function wakeup support
USB3.2 spec section 9.2.5.4 quotes that a function may signal that
it wants to exit from Function Suspend by sending a Function
Wake Notification to the host if it is enabled for function
remote wakeup. Add an api in composite layer that can be used
by the function drivers to support this feature. Also expose
a gadget op so that composite layer can trigger a wakeup request
to the UDC driver.

Change-Id: Iab3153da8998b60d8f1a77f769c7167033dced93
Signed-off-by: Elson Roy Serrao <quic_eserrao@quicinc.com>