include: linux: pcs: Compatibility for Qualcomm PCS
Add an IO address to the dw_xpcs struct in order for
Qualcomm's new xpcs_read and xpcs_write APIs to work
without MDIO bus. Direct memory reads and writes are used
to update PCS control and status registers.
Add a macro for USXGMII interfaces to work with C37
Autonegotiation.

Change-Id: I9bc20bff6d887f928f5fb48ce05ef06f7249c4da
Signed-off-by: Manuel Gomez <quic_gomez@quicinc.com>