sched: walt: handle symmetric 32/64 bit system in halt
Systems with symmetric ELF architecture (all cores supporting 32bit or none
of the cores supporting 32bit), there is a possibility of having all the
cores in the system in either halted or offline state and thus leaving
no active cpu for scheduling.
Handle this usecase by updating halt handling to ensure first possible core
in the system is never offlined or halted.

Change-Id: I8caa9f352ac08eb9946a1bd83bfe58be3ef0d966
Signed-off-by: Ashay Jaiswal <quic_ashayj@quicinc.com>