ARM: Kconfig: Enable TRACING
The CONFIG_FTRACE might not be selected when kernel is build for
sdxbaagha SoC and, as result TRACEPOINTS option will not be selected
also. TRACEPOINTS is selected by TRACING and TRACEPOINTS is used for
vendor hooks. So the functions which need vendor hooks will be broken,
such as LPM, Remoteproc and so on.

Hence, fix it by selecting TRACING options for sdxbaagha SoCs
always.

Change-Id: If3f931f2aed6c0c5a456166b4e3f856a92a8e38f
Signed-off-by: Lijuan Gao <quic_lijuang@quicinc.com>