mtd: msm_qpic_nand: fix for boot partition with suffix
In some devices, like mirror system we have boot_a/_b partition
not just boot partition.
So, one of the device with bad MCP memory part was crashing.

Also, the serial logs reports wrong parameter configuration:
[    2.778373] msm_nand_flash_onfi_probe: msm_nand_flash_onfi_probe: no apps partition found in smem
[    2.786633] msm_nand_scan: missing page size info - extract from NAND ID
[    2.795513] msm_nand_scan: NAND Id: 0x6d0a32c Buswidth: 8Bits Density: 1024 MByte
[    2.802280] msm_nand_scan: pagesize: 4096 Erasesize: 65536 oobsize: 128 (in Bytes)
Here, the Erasesize is reported wrongly.

To fix this, replace strcmp with strncmp to compare with all combination of
boot_x name of partition.

Change-Id: Iad993b17a558e0f02e62c2bfab050fd136894d31
Signed-off-by: Pintu Kumar <quic_pintu@quicinc.com>