msm: mhi-dev: Move sm_init ahead of pcie events registration
When the endpoint device undergoes a warm-boot, the config space and
BME of VF instances are kept intact by the host.

Due to this reason while enumerating MHI, when it registers for PCIe
link-up event of any VF instance, PCIe driver triggers the link-up
event (for that VF) immediately.

To handle this event, mhi state-machine (sm) initialization should have
completed by that time. So moving mhi_sm_init ahead of PCIe events
registration in both init and re-init paths.

Without this, the driver fails to catch link-up event for VFs and
thus fail to initialize MHI VFs in the warm-boot path.

Change-Id: I89ddec34580ea699ac8682af1410c718bebd2382
Signed-off-by: Subramanian Ananthanarayanan <quic_skananth@quicinc.com>
Signed-off-by: Veerabhadrarao Badiganti <quic_vbadigan@quicinc.com>