msm: mhi-dev: Update status register to reflect right state for VFs
Unlike physical function (PF), for virtual functions (VFs) MHI state
field in MHISTATUS register won't be set to reset state by the
bootloader in the warm-reboot path.

So update the status register to reflect the right state for VFs.

Without this change, the driver fails to set VFs to the ready state
in the warm-boot path as it can't move the mhi-state from the error
state to the reset state.

Change-Id: I4fbed53ed2dfb15ae975c7dc58ae8f518184ea78
Signed-off-by: Veerabhadrarao Badiganti <quic_vbadigan@quicinc.com>