usb: gadget: Properly configure the device for remote wakeup
The wakeup bit in the bmAttributes field indicates whether the device
is configured for remote wakeup. But this field should be allowed to
set only if the UDC supports such wakeup mechanism. So configure this
field based on UDC capability. Also inform the UDC whether the device
is configured for remote wakeup by implementing a gadget op.

Change-Id: I1ead53afa469697cd9ae9ac9d1ea172db4978843
Signed-off-by: Elson Roy Serrao <quic_eserrao@quicinc.com>