i2c: i2c-msm-geni: Add a new error code for Spurious irq
Currently in irq handler, there is a check for current message
pointer with NULL. We are not updating any error code if current
message ponits to NULL. Due to this, we are going for
i2c second xfer(READ) without i2c first xfer (WRITE) success on bus.
To solve that updating an error code as spurious interrupt received
when current message ponits to NULL and also that we don't serve
read if write is failing.

Change-Id: I78fe370a8d80010a9f55fda0e69d50670f82d06b
Signed-off-by: Viken Dadhaniya <quic_vdadhani@quicinc.com>