gunyah: dbl: fix the might sleep in an atomic context
When call gh_dbl_validate_params with flag GH_DBL_NONBLOCK
in an atomic context, it will call sleep in wait_event_interruptible
and cause the kernel error. So we can return directly to avoid
the sleep in the atomic context, when cap_id is not invalid and
the flag is GH_DBL_NONBLOCK from the caller.

Change-Id: I63994e86e457fbbdf27e5c20fb033b98c3e66956
Signed-off-by: Peng Yang <quic_penyan@quicinc.com>
Signed-off-by: Arun Prakash <quic_app@quicinc.com>