drivers: dcvs: memlat: fix probe failure when all CPUs in MON are disabled
There can be a scenario where all the CPUs in a MON are disabled
in which case num_inited_mons of that group will not get incremented for
that monitor. This results in memlat probe failure as the number of MONs
that are read from device tree do not match with the actual number of
monitors that are initialized during probe. Fix the issue by decrementing
the num_inited_mons for the corresponding monitor when all the CPUs in it
are disabled.

Change-Id: I07f2e8c2e5c81f1d444fb88545a16d411faaee4b
Signed-off-by: Vandana Gupta <quic_vandgupt@quicinc.com>