serial: msm_geni_serial: Handle uart alias line id correctly
Current uart driver is not handling line id of
uart alias instances and non-alias instances correctly.
Due to this crash is observed when ttyHS1 is probed
before ttyHS0.

Updated the logic to handle uart alias line id
correctly, and also return error if probe is called
with same alias twice.

Change-Id: Ia441e275dade1a32874225ebaa588648e812f661
Signed-off-by: Zahir Shabbir Khan <quic_zshabbir@quicinc.com>