arm64: defconfig: sdxpinn: Enable msm_performance
This commit enables the msm performance kernel module for sdxpinn
target. The perf service requires this module in order to vote for
the maximum and minimum CPU frequencies, among other things.
So by enabling this module, the perf service will be able to more
efficiently manage the CPU frequencies and improve system performance.

Change-Id: I90ae1b100c73c99981a947adab4a8cbd9627afc2
Signed-off-by: Rajat Asthana <quic_rasthana@quicinc.com>