net: stmmac: Seemless switch between HW path to SW path
When IPA offload path is disconnected we do not have a firm solution
to seemlessly move to Sw path, there are several dangling pointers,
several memory leaks and most of all the whole system is in a panic
state. This change will seemlessly solve all the issue and making sure
that both HW and SW path work seemlessly.

Change-Id: I53b3b8c75815a7c431187b7747cf75b0d7e029fd
Signed-off-by: Abhishek Chauhan <quic_abchauha@quicinc.com>