msm: pcie: Dumping DBI registers of AER before the suspending link
There can exist a possible race condition between dumping the
dbi registers as a part of AER and immediate link suspend request
from the client. Link suspend operation turns off all the voted clks
and dumping of the DBI space registers requires the clks to be on.
To avoid this, introduce a  mutex lock in "msm_handle_error_source"
api to collect the DBI reg dump before triggering the suspend
operation.

Change-Id: I1c92b440e8e064d30df57680915ef349f5b51600
Signed-off-by: Naga Rashmi Ayiluri <quic_nayiluri@quicinc.com>