msm: mhi_dev: avoid race between mhi_hwc_init and mhi_hwc_cb
Change is to add mhi_lock to mhi_dev_enable WQ to prevent the
race condition where mhi_hwc_cb() is called prior to mhi_dma_init()
API being fully completed as part of mhi_hwc_init() call.

This causes M0 event bebing sent to the host prior to db_routing()
being programmed thus causing an SMMU fault with the address that was
programmed as part of init prior to the WARM boot.

Change-Id: I96b906aab869d211f373c0cbd1e81faa292cf4cc
Signed-off-by: Subramanian Ananthanarayanan <quic_skananth@quicinc.com>