defconfig: Configure QRTR Node ID for local processor
Configure QRTR Node ID for local processor.
The node ID published to other nodes within the system.
This value can be overridden by the name service application.

When the node id configured as 1 the services available in mdm
QRTR not forwarding because LinuxHost/MSM QRTR node id and mdm
node id are same. So the node id value changed to 2 to
forward all available services on MDM to MSM.

Change-Id: Id11d696f67919d507d807eae423eb51819204f3a
Signed-off-by: Sivaji Boddupilli <quic_boddupil@quicinc.com>