uart: msm_geni_serial: Add shutdown callback for uart
If the hardware is still accessing memory after SMMU translation
is disabled (as part of smmu shutdown callback), then the IOVAs
(I/O virtual address) which it was using will go on the bus
as the physical addresses which will result in unknown crashes
like NoC/interconnect errors.

So, implement shutdown callback by removing serial driver before
SMMU is disabled.

Change-Id: I5398cf7efbb15a04bcb7259f7580e8d88e4647be
Signed-off-by: Prasanna S <quic_prass@quicinc.com>