msm: mhi_dev: Remove additional check for event_rd_dma
The argument validation in mhi_dev_event_rd_offset_completion_cb for
event_rd_dma is returning from the above API. With changes to support
the host without MSI capability, the clearing of cmd_req_buffers is
handled from this same API. Due to the above issue, the cmd_req_buffers
are not getting cleared and running out. Hence, removing this additional
check for this parameter as there is a conditional execution with the
same parameter in mhi_dev_event_rd_offset_completion_cb.

Change-Id: I4c26850e076271ec44555d0564ce1cf16aa485c0
Signed-off-by: Mohana Basava Tejesh Reddy Mareddy <quic_mmareddy@quicinc.com>