usb: phy: phy-msm-snps-eusb2: Toggle VIOCTL_EUD_DETECT during disconnect
When operating in HSUSB with EUD enabled, if the user physically
disconnects the cable, there is no disconnect event being generated.  This
does not happen when the disconnect is done over tools, such as Alpaca.
The suggested HW sequence is to toggle the EUD detect register, which in
turn places the EUD and USB PHY in reset, which allows the disconnect event
to be generated.  Only execute the sequence when EUD is active.

Change-Id: Ia43e5165d4852355c2e47a55cd271c671ae41872
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>