Revert "remoteproc: qcom: pas: Rename subsystem minidump elf name"
The reason of the revert of this patch is because this
patch does not uses 'inline' in its implementation but
always assume 'enabled' in its implementation which
allocate a buffer of bigger size and that can result
in OOM issue for low memory target like sdxpinn.

The inline implementation of this patch was needing a
change in core kernel and due to GKI constraints it was
not done during 5.10 target time.

Everything works functionally even after this revert and
revert is only being done for 5.15-c1 branch, the only
thing it effects is the name differentiation of the elf
name for minidump and full dump.
This reverts commit 798e5a838a0fef55afbba91442416fc2b84c8885.

Change-Id: Ibc80ccd64b68c28500713e5231a09779ac3fd5ce
Signed-off-by: Rohit Agarwal <quic_rohiagar@quicinc.com>