integrity: Define QCOM IMA appraisal measure rules
1. Add configs for QCOM specific IMA Feature.

2. Define IMA appraisal measure rules based on uuid.
The measure and appraise rules are defined in kernel
driver itself instead of initramfs or loaded from
user space during early boot.

When a ubi image is created, uuid is generated which is
a random number every time. This image when flashed in
a partition helps in identifying that partition based
on uuid. Because we need to define a UUID in IMA rule,
we are using pre-defined UUIDs for rootfs and data
partitions.

Change-Id: I4b37d59b6900efbfdb7ba0c79b13498b777202cf
Signed-off-by: Manoj Choudhary <quic_cmanoj@quicinc.com>