integrity: disable creating IMA sysfs entries
Disable creating sysfs entries for ima policy.
The rules to measure and appraise files are now embedded
in driver file ima-policy.c itself. Therefore there is no
need for sysfs files that are required to send new rules
to kernel. This commit removes call to the logic to create
those files. Additionally, it increases security a bit also.

Change-Id: I524bbca6f32e0f72af564ab1dea4c204715dfead
Signed-off-by: Manoj Choudhary <quic_cmanoj@quicinc.com>