usb: dwc3: msm: Cancel worksync quickly during shutdown
In some cases it is observed that during reboot, there is an instance
of sm work that was queued and it races with platform remove callback
invoked during reboot. The following trace occurs in dmesg:

69.872579:    __queue_work+0x1c0/0x570
69.877058:    queue_delayed_work_on+0xd0/0x14c
69.882245:    dwc3_ext_event_notify+0x2c8/0x4c0
69.887519:    dwc3_msm_set_role+0x234/0x254
69.892439:    dwc3_msm_usb_role_switch_set_role+0x18/0x28
69.898611:    usb_role_switch_set_role+0x4c/0x8c
69.903982:    ucsi_handle_connector_change+0x348/0x638
69.909880:    process_one_work+0x214/0x484
69.914712:    worker_thread+0x248/0x520
69.919273:    kthread+0x168/0x1b0
69.923311:    ret_from_fork+0x10/0x20

and at the instant of reboot, we see the following signature:

69.441671:   sysrq: Emergency Remount R/O
69.446872:   Emergency Remount complete
69.487398:   android_work: sent uevent USB_STATE=DISCONNECTED
69.555650:   platform a600000.dwc3: Removing from iommu group 14

Also from dumps indicate that platform remove was successful but still
the delayed work wasn't yet completed successfully.

So, cancel sm_work and destroy it soon after de-registration to avoid
any instance of that work queue racing with remove callback.

Change-Id: I3998b6f1dac84e44177eb3ba78103967efd019a6
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>