firmware: Add ARM SCMI virtio backend implementation
Add implementation for SCMI virtio backend, for handling SCMI
requests received over virtio transport. All SCMI
requests from a guest are associated with a unique
client handle. The SCMI request packet, received from
the SCMI virtio layer is unpacked and depending on the
message id, is forwarded to the correct protocol, if
that protocol is implemented. The response packet is
packed with the payload returned by the protocol layer,
and message header is added to it.

Available protocols are specified using child device tree
nodes of the parent scmi virtio backend node. Current
patch only implements the base protocol.

Notifications and delayed response are not supported at this
point.

As all incoming SCMI requests from a guest are associated
with a unique handle, which is also forwarded to the
protocol layer; on channel close call for a guest, protocols
can do the required cleanup of the resources allocated for
that particular guest.

Change-Id: I60aceb6e7818c39eaed3ba6e5d407e6a9aa40aa7
Co-developed-by: Srinivas Ramana <quic_sramana@quicinc.com>
Signed-off-by: Srinivas Ramana <quic_sramana@quicinc.com>
Signed-off-by: Neeraj Upadhyay <quic_neeraju@quicinc.com>
Signed-off-by: Johnathan Cobden-Nolan <quic_jcobdenn@quicinc.com>