spi: spi-msm-geni: Add cleanup support in SSR path for SVM
Sometimes if data transfer is ongoing and VM gets a crash due
to unknown reason then there is a possibility of garbage data
present in Tx/Rx FIFO's. Due to this next transfer might be fail.

To avoid this, GENI need to be reset in probe time if it's
in bad state.

Change-Id: I18114119d36aa1aa68b758985f77c00c706ea238
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>