msm: mhi_dev: Wake host on IPA DMA fun ops call if link is disabled
MHI_MMIO_INIT is done after PCIe link is enabled and IPA DMA is
available. Consider the ADB reboot in D3hot/D3cold cases. After reboot,
Link training and enumeration with host is completed first. But at this
time IPA DMA fun ops call is not yet received. IPA DMA fun ops call is
received after a delay, at around 20 secs of HLOS start in sdxbaagha and
sdxpinn M2 endpoints. Since Host is unaware of device state it sends
D3hot and D3cold events on inactivity timeout. IPA DMA fun ops is seen
to be received after this. In the current implementation, MHI_MMIO_INIT
is called directly leading to NOC error on accessing MHI status register
since device is in D3cold state.

To handle the above scenario where Link is not enabled when IPA DMA fun
ops is received, added a check to see the link status and wake up the
host if Link is not enabled. Also, added registering for link up event
if it is not registered as part of MHI probe, so that linkup event is
not missed after host wake up.

cherry-pick 4778603/4

Change-Id: I0eaec2222cf59a1133a6877fbb97e4cc47dc90b4
Signed-off-by: Sai Chaitanya Kaveti <quic_skaveti@quicinc.com>