pci: msm: Correct the logic of programming preset values
Ideally driver need to find the extended capability of Physical
Layer 16.0 GT/s Extended Capability for Gen4 and Secondary PCI
Express Extended Capability for Gen3 instead of hardcording the
offset as offsets change from target to target.

And at current driver is programing only for 2 lanes for GEN3 and
4 lanes for GEN4.

So, updated the logic for both cases.

Change-Id: I1fc0bae41434557048a4566eb36a97737da52f75
Signed-off-by: Krishna chaitanya chundru <quic_krichai@quicinc.com>