msm: mhi_dev: Allocating memory for channels during runtime
Allocating memory for channels during runtime by making an array of
pointers instead of assigning memory for 128 channels by default, to
optimize memory utilization. Memory is assigned to the channels only
when they are opened. This mempory is not released and held forever even
after the channels are closed for ensuring easier debugging.

For 64bit architecutre based SoCs, Previously the memory allocation per
channels was 272 bytes. So for 128 channels (272*128)=34,816bytes~34Kb was
allocated, of which only a fraction was used. Now memory allocated per
channel pointer is 8 bytes. So for 128 channels(128*8)=1kb is allocated.
Total memory used now is 1kb+((n*272)/1024)kb where n is no of
opened channels.

For 32bit architecutre based SoCs, Previously the memory allocation per
channels was 144 bytes. So for 128 channels (144*128)=18,432bytes~18Kb was
allocated, of which only a fraction was used. Now memory allocated per
channel pointer is 4 bytes. So for 128 channels (128*4) =512bytes~0.5Kb is
allocated. Total memory used now is 0.5kb+((n*144)/1024)kb where n is no of
opened channels.

Change-Id: I0e83520aa1f4f9d9f66903c0efdaaef653dba97e
Signed-off-by: Shivangi Yadav <quic_shivyada@quicinc.com>