net: stmmac: reset global L3/L4 filter variables
Make sure we deinit all the variable with respect to
L3 L4 and PTP over UDP filtering when the local link goes down.
When the Link comes back up we do DMA reset and everything is
cleanly installed from the beginning.

All global variables are reset as part of this change.

PTP over UDP filter should be only installed for port 0 as of now.

Change-Id: Ic90487465a7ac5144266dbe85f8dadb504f4ddd6
Signed-off-by: Abhishek Chauhan <quic_abchauha@quicinc.com>