rpmsg: glink: Print signal value in hex
Print signal value in hex to identify set bits easily.
When MSB is set while adding signal bits such as CTS/RTS/DTS
the computed value is in the form of negitive. if we print
value with %d format specifier user can interpret it as some
incorrect value.

This patch will print the value in hex format to interpret
signal value easily.

Change-Id: Iaaba9fcb980cc9ecfddd1c29e8c47330a279a130
Signed-off-by: Sivaji Boddupilli <quic_boddupil@quicinc.com>