msm: mhi_dev: Fix for inbound transfer size greater than 4k
The inbound transfers with the size greater than that of the
TRE length configured by the host should be split based on the
TRE size as per the MHI spec. For instance, if the TRE len is 4k
and a transfer of size 6k has to be made from device to host, it
is expected to transfer with 4k and 2k sizes. The implementation
is in-place in mhi_dev_write_channel to accommodate this feature.
However, the comparison check for transfer size and TRE len is
blocking it by returning from that point if transfer size is
greater. Hence, removing the condition to proceed with the
implementation to achieve splitting the transfer as expected.

Change-Id: I0ae4d7dec0ea468b9adaff0b12a67d4b2a72dfd1
Signed-off-by: Mohana Basava Tejesh Reddy Mareddy <quic_mmareddy@quicinc.com>