defconfig: Enable CONFIG_SOFT_WATCHDOG for cinder
Currently cinder has a h/w wdog driver that manages pet from kernel. It
is required that user space should have the ability to pet so as to
monitor health of key system service and be able to restart the system
if it is hung.

Enable CONFIG_SOFT_WATCHDOG to provide a s/w wdog driver that can cover
the user space requirement while the current h/w wdog driver continues
to work as it is.

Change-Id: I821033427f11252f0da6b80c8cef74f1e9111c86
Signed-off-by: Khaja Hussain Shaik Khaji <quic_kshaikkh@quicinc.com>