dwc3-msm-core: Avoid regdump if in LPM during smmu fault
commit 1084114ec372 ("usb: dwc3: dwc3-msm-core: Add smmmu
fault handler and dump the registers") was introduced
recently which helps in dumping the dwc3 & qscratch registers
during smmu fault scenario.

However, the changes didn't take into account that if dwc3/mdwc is in
LPM and a kernel panic/smmu fault is encountered, in those cases also
the functions will try to dump the registers being unaware of
the state of the device which will result in a NOC error due to
unclocked access.

Fix this by avoiding register dump if we are in LPM since it will be
evident that usb is not in use.

Change-Id: If645cdab8d3d87e8d5dcbebe4ac6c527087a2a8f
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>
Signed-off-by: Sriram Dash <quic_sriramd@quicinc.com>