usb: f_gsi: Handle pending notification request during suspend
If host suspended bus without receiving notification request then
initiate remote-wakeup. As driver won't be able to do it later since
notification request is already queued. If remote wakeup is not
allowed, dequeue that request as we are in some incomplete state and
this request is of no use now and most likely we would initiate a
protocol disconnect during suspend. Keeping the request queued might
also prevent the underlying UDC from entering runtime suspend.

Change-Id: I464d9ab85b07c5ca0b7e52c56e35125cfed5bf1a
Signed-off-by: Elson Roy Serrao <quic_eserrao@quicinc.com>