ALSA: pcm - add support for hostless audio
Allow some PCM devices to be hostless, i.e. there is no
PCM data transferred to or from the host CPU. This can
be used to minimise power on system since the CPU can
idle/sleep during the PCM device operation (e.g. a phone
call where the DAI is between a MODEM and DSP).

Signed-off-by: Kumar Anurag Singh <quic_kumaranu@quicinc.com>
Change-Id: Ib2bdb04e54a62ec7b9e26ec0eba7f3975ab7b610