phy: msm: snps: eusb2: Update the eusb2 init sequence
Currently the driver is following only one init sequence for
all chipsets which seem to be wrong as the latest HPG revision
describes additional changes in the sequence which is missing.

This change add the missing register writes into the eusb2 init
sequence to make sure of the proper functionality.

Description:
USB_PHY_CFG0: setting to 0x00 at the end of init sequence.

Change-Id: Id35e59ba25aed5a489cfe918160be2ea29ea06ea
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>