msm: ep_pcie: Provision to configure PS_HOLD to hard reset
To support host warm reboot scenario, this change provides the
provision to configure PS_HOLD to hard reset. When host issues
D3hot to device, the PMIC gets configured to hard reset. Since
PERST coupling is enabled, when host issues D3cold this causes
the device to hard reset along with the pcie controller as
part of PERST coupling.

Change-Id: Id3c04526a38ca45e2716fe132004b20afa7f3dd5
Signed-off-by: Subramanian Ananthanarayanan <quic_skananth@quicinc.com>