pci: msm: Add a check for lpi_enable before msi config access
For LPI based interrupts, IRQ handler passes the IRQ directly
to the client. So, no need of pcie driver(pci-msm-msi.c) for LPI
based interrupts.

Add a check for lpi_enable before msi config access to MSI
register at the time of link enumeration as msi config access
is not applicable to LPIs.

Change-Id: Ice55eb5df9f52bab8b4a029ed55236ecf1766766
Signed-off-by: Paras Sharma <quic_parass@quicinc.com>