msm: ep_pcie: Clear the RO bit irrespective of the flag "use_iatu_msi"
With <3cdec24cf1627>(ep_pcie: core: Add support for SRIOV)
"use_iatu_msi" flag was introduced and clearing RO bit logic
was put under this flag, but we should create this iATU
irrespective of use_iatu_msi flag.

Whenever EP tries to generate PARF based MSI for Completion Event
on QMI channels(which follows SYNC path for data transfers)
we were seeing that RC is dropping this Memory Write.

If we don't create RO bit clearing iATU then interrupt
associated with the MSI TLP will not be generated at
GIC Interrupt controller on RC.

Hence moving the creation logic for the iATU for the
"RO bit clearing" out of the flag use-iatu-msi,
so that the RO bit is cleared before sending the MSI TLP.

Change-Id: Ieac2fcb6077bfeb91eb4a06a560ccd7c0f9fcf6c
Signed-off-by: Naga Rashmi Ayiluri <quic_nayiluri@quicinc.com>