driver : hdcp_qseecom: Retry HDCP key verify when Qseecomd down
HDCP has a dependency on Qseecomd for HDCP key
verification and other SCM calls. Qseecomd takes
a lot of time to be available during boot because
of which HDCP init fails.
Adding retry mechanism only when there is hdcp key
verification failure due to Qseecomd not starting
in time. For other error codes, like HDCP key not
programmed, there will be no retry.
If Qseecomd doesn’t come up within maximum retries,
code will stop retrying and hdcp key verification
will fail. For each retry, there is a sleep of 100ms.
The number of retries can be changed from the dtsi file.
By default, we are allowing 40 retries but it can be
changed. 40 retries are based on the qseecomd KPI which
can be anywhere between 2.9 sec to 3.6 sec.

Change-Id: Ieb6ed6c25664ef8609602711f664cab045ce6979
Signed-off-by: Shubham Dhiman <quic_shubhamd@quicinc.com>
Signed-off-by: jueqi wang <quic_jueqwang@quicinc.com>