net: stmmac: resource leak when driver is removed
Regulators are not disabled when driver is removed due to a coding
error in the remove mac error recovery module.
We are seeing that driver remove is never successful and resulting
in future binding of driver.
The coding error is fixed now. All bind and unbind operation will be
successful from now on.

Change-Id: I48ba67676f2c2e72fe01241749c0d9be3dc4136a
Signed-off-by: Abhishek Chauhan <quic_abchauha@quicinc.com>