mtd: msm_qpic_nand: Fix Overwriting MSM_NAND_CTRL register
Do not configure BAM_MODE_Enable if it is already enabled. This
bit will be configured in bootloader.

Also in current code, as part of msm_nand_enable_dma function
BAM_MODE_Enable is being configured and it is overwriting the
entire register contents. Read the MSM_NAND_CTRL and then append
the BAM_MODE_Enable with out overwriting the content.

Change-Id: I39c1c7412dfa3f7efda04ccc0d5b5475f24d6db0
Signed-off-by: Naveen Kumar Goud Arepalli <quic_narepall@quicinc.com>