usb: gadget: f_cdev: Avoid requeueing the USB requests from complete callback
Commit 51ada3a9945a ("usb: gadget: cser: Requeue OUT USB requests
if ECONNRESET is received") requeues the USB requests from the
complete callback if gadget layer gives back the requests with
status set to -EPIPE or -ECONNRESET. But consider the following
scenario happening during f_cdev_open() or  usb_cser_set_alt(),

  usb_ep_queue			--> starts recursing
  usb_cser_read_complete	--> req->complete()
  dwc3_gadget_giveback
  ep_cleanup_cancelled_requests	--> sets status as -ECONNRESET
  __dwc3_gadget_kick_transfer	--> gadget_send_ep_cmd fails and
				    req moved to cancelled list
  usb_ep_queue
  usb_cser_start_rx

This recursion will eventually lead to memory exhaustion and
page faults. Avoid this by not requeueing the USB request from
complete callback in case of ECONNRESET, but move the failed
USB request to read_pool which will be fetched by the ongoing
cser_start_rx. So if we hit the start_transfer failure, then
the following is expected to happen.

- cser_start_rx()		--> takes spin_lock(port), and
|	  |			    process next request from
|	  |			    read_pool if its not empty.
|	  |
|    	  |			--> release spinlock(port) and
|	return			    returns back to start_rx()
|	list_add(req)
|     cser_read_complete	--> acquires spinlock(port)
|     gadget_giveback		--> sets status = -ECONNRESET
|     gadget_send_ep_cmd
|   usb_ep_queue		--> outside spin_unlock(port)
->usb_cser_start_rx()

Meanwhile -EPIPE is returned when host clears EP_HALT, and it
won't cause recursion even if requeue the USB request from
read_complete() callback.

Change-Id: I37fa264d8a7b24948433d770f00ade4da0366d44
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>