UPSTREAM: usb: dwc3: Poll CMDACT after issuing EndXfer command
Currently for the controller revisions that support GUCTL[14]:
Rst_actbitlater bit, we don't wait for 1ms after issuing the
DEPENDXFER command. And since we proceed with the assumption
that EndXfer got completed, and giveback the requests, which
eventually frees these requests. If EndXfer command doesn't
get issued successfully, and the controller still continues
to access requests from the previous UpdateXfers, this leads
to SMMU faults.

DWC3 databook specifies that CMDACT bit can be polled to check
completion of the EndXfer

Section 3.2.2.7 Command 8: End Transfer (DEPENDXFER)
Note: If GUCTL2[Rst_actbitlater] is set, Software can poll the
completion of the End  Transfer command by polling the command
active bit to be cleared to 0.

TODO:   Update commit text
	Try reusing polling from send_gadget_ep_cmd
	determine the rety timeout

Change-Id: Idf46e9d38a1da394dc0e4e158175f03580b9da34
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>