msm: mhi_dev: Support channel error state change notification
Change is to support Q6 SSR by allowing the notification provided
by DMA ECPRI/IPA for SSR by indicating CH_ERROR state event to host.
Channel state is updated to ERROR state in channel context by
ECPRI/IPA for all HW channels that are involved as par of Q6 SSR.
ECPRI/IPA driver then provides a CB to MHI driver. As part of this,
we send an event that enables the host to look into channel context
array of all channel and process the ones that are in ERROR state.

			 ---------------------	 	-----------
			| Host processes      |		| MHI sends |
			| event and indicate  |		| CH_ERROR  |
			| to clients who then |  <--	| event to  |
			| cleanup and restart |		| host	    |
			| the channels.	      |	 	-----------
			 ---------------------			^
								|
 -------	 ---------------------------------	-----------
| Q6 SSR | -> 	| ECPRI updated CH's ctxt to ERROR| -> | CB to MHI |
 -------	 ---------------------------------	-----------

This is currently supported only for H/W channels.

Change-Id: I77901b636fb6f63c08316623c27cb60e12e37f53
Signed-off-by: Subramanian Ananthanarayanan <quic_skananth@quicinc.com>