msm: ep_pcie: Additional check to make sure linkup happens in L2
This additional check is to make sure, link re-enumeration happens
in HLOS if link is in L2.

For targets which support retention, some of the PCIe registers were
not getting cleared as part of GDSC power down and power up. If
retention is not latched in HW, PCIe controller also gets reset as part
of GDSC reset sequence. This makes sure there are no stale value in PCIe
registers. For targets where retention is enabled, we can't rely on XMLH
linkup register to decide if Link retraining is needed or not. As part
of this change we are adding extra flag to check current link state. If
link is in L2, link retraining will be done, even if status register
shows link up.

Change-Id: Ib20e48d798d3044afefa1b562532e9831f708332
Signed-off-by: Pallavi Singh <quic_pallsing@quicinc.com>