spi: spi-msm-geni: Configurable vote for cpu, geni and ddr
The SPI driver initializes resources with a fixed
clock vote for the CPU, Geni, and DDR. However,
these values vary from target to target.

To address this, read the clock vote values from
the DTSI file and configure and initialize the
resources accordingly. If the resources are not
explicitly mentioned in the DTSI, they will be
initialized with default values.

Change-Id: I418f65200d66536fff53a7170e3f942d7d430fd7
Signed-off-by: Mehul Raninga <quic_mraninga@quicinc.com>