usb: dwc3: dwc3-msm-ops: Stop pending GSI transfers on reset
In order to ensure that GSI transfers are stopped on a bus reset
condition, do not rely on f_gsi issuing the EP op to do it. The
f_gsi driver would handle this asynchronously within the IPA SM
work, which could run later in the timeline when the bus reset
routine has already completed.  Utilize the DWC3 reset kretprobe
to issue the endxfer command for any pending GSI endpoints.

Change-Id: Iad8ae1327c440eaab47606ec380dfbac77f24fc3
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>