mmc: sdhci-msm: Update clock doubling logic correctly for HS400 mode
As per Hardware programming guide sequence for entering HS400 mode clock
should not exceed 52Mhz when switching from HS200 to HS400 mode post
tuning and before sending CMD13 to the device.

But in the current design since the timing is set to HS400 first which
is a DDR mode we end up doubling the clock to 104Mhz instead of 52Mhz
in transition phases which is a spec violation and leads to CRC
error for the next CMD13 being sent or sometimes CMD timeouts.

Avoid doubling the clock in between any transition phases and double
only when 200Mhz needs to be supplied to the card for HS400 mode.

Change-Id: Ia6db0955549bc7aab8fa0ef55ce255e7f3dacc70
Signed-off-by: Sarthak Garg <quic_sartgarg@quicinc.com>