net: qrtr: bpf_service: Fix rcu logic
RCU locks still require the updating of the object to be protected by
locks or semaphores. Add a new lock to ensure writes to the bpf_filter
pointer cannot race against each other.

Expand the rcu_read_locks to ensure they are held during all places
where the bpf_filter pointer is being used. This will ensure that
references are properly accounted for and can be synchronized on before
attempting to free the filter in the detach ioctl.

Change-Id: Ib4a9d6e3b44fa6253164c50f52c627e64941baea
Signed-off-by: Chris Lew <quic_clew@quicinc.com>